requirements.txt
notebook
opencv-python
joblib==1.4.2
biopython==1.80
keras==2.10.0
Keras-Preprocessing==1.1.2
numpy==1.23.5
pandas==1.5.2
plotly==5.11.0
scikit-learn==1.1.3
scipy==1.9.3
shap==0.40.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.28.0
tensorstore==0.1.41
umap-learn==0.5.6
xgboost==1.4.2
matplotlib==3.6.2